requirements.txt
aiohttp==3.8.1
attrs==21.2.0